-- imported fixes from the commercial version (EvoStream http://www.evostream.com)
git-svn-id: https://svn.rtmpd.com/crtmpserver/trunk@776 01689507-d990-df11-9d7d-0017a411d416

+++ sources/thelib/include/mediaformats/mp4/atomuuid.h
@@ -27,7 +27,7 @@ class AtomUUID
 private:
 	Variant _metadata;
 public:
-	AtomUUID(MP4Document *pDocument, uint32_t type, uint32_t size, uint32_t start);
+	AtomUUID(MP4Document *pDocument, uint32_t type, uint64_t size, uint64_t start);
 	virtual ~AtomUUID();
 
 	virtual bool Read();
